+++ paddlespeech/audio/backends/soundfile_backend.py
@@ -23,11 +23,11 @@ import soundfile as sf
 from scipy.io import wavfile
 
 from ..utils import ParameterError
+from ..utils import depth_convert
 
 __all__ = [
     'resample',
     'to_mono',
-    'depth_convert',
     'normalize',
     'save',
     'load',
@@ -115,78 +115,6 @@ def to_mono(y: np.ndarray, merge_type: str='average') -> np.ndarray:
     else:
         raise ParameterError(f'Unsupported dtype: {y.dtype}')
     return y_out
-
-
-def _safe_cast(y: np.ndarray, dtype: Union[type, str]) -> np.ndarray:
-    """Data type casting in a safe way, i.e., prevent overflow or underflow.
-
-    Args:
-        y (np.ndarray): Input waveform array in 1D or 2D.
-        dtype (Union[type, str]): Data type of waveform.
-
-    Returns:
-        np.ndarray: `y` after safe casting.
-    """
-    if 'float' in str(y.dtype):
-        return np.clip(y, np.finfo(dtype).min,
-                       np.finfo(dtype).max).astype(dtype)
-    else:
-        return np.clip(y, np.iinfo(dtype).min,
-                       np.iinfo(dtype).max).astype(dtype)
-
-
-def depth_convert(y: np.ndarray, dtype: Union[type, str]) -> np.ndarray:
-    """Convert audio array to target dtype safely. This function convert audio waveform to a target dtype, with addition steps of
-    preventing overflow/underflow and preserving audio range.
-
-    Args:
-        y (np.ndarray): Input waveform array in 1D or 2D.
-        dtype (Union[type, str]): Data type of waveform.
-
-    Returns:
-        np.ndarray: `y` after safe casting.
-    """
-
-    SUPPORT_DTYPE = ['int16', 'int8', 'float32', 'float64']
-    if y.dtype not in SUPPORT_DTYPE:
-        raise ParameterError(
-            'Unsupported audio dtype, '
-            f'y.dtype is {y.dtype}, supported dtypes are {SUPPORT_DTYPE}')
-
-    if dtype not in SUPPORT_DTYPE:
-        raise ParameterError(
-            'Unsupported audio dtype, '
-            f'target dtype  is {dtype}, supported dtypes are {SUPPORT_DTYPE}')
-
-    if dtype == y.dtype:
-        return y
-
-    if dtype == 'float64' and y.dtype == 'float32':
-        return _safe_cast(y, dtype)
-    if dtype == 'float32' and y.dtype == 'float64':
-        return _safe_cast(y, dtype)
-
-    if dtype == 'int16' or dtype == 'int8':
-        if y.dtype in ['float64', 'float32']:
-            factor = np.iinfo(dtype).max
-            y = np.clip(y * factor, np.iinfo(dtype).min,
-                        np.iinfo(dtype).max).astype(dtype)
-            y = y.astype(dtype)
-        else:
-            if dtype == 'int16' and y.dtype == 'int8':
-                factor = np.iinfo('int16').max / np.iinfo('int8').max - EPS
-                y = y.astype('float32') * factor
-                y = y.astype('int16')
-
-            else:  # dtype == 'int8' and y.dtype=='int16':
-                y = y.astype('int32') * np.iinfo('int8').max / \
-                    np.iinfo('int16').max
-                y = y.astype('int8')
-
-    if dtype in ['float32', 'float64']:
-        org_dtype = y.dtype
-        y = y.astype(dtype) / np.iinfo(org_dtype).max
-    return y
 
 
 def sound_file_load(file: os.PathLike,
@@ -323,3 +251,7 @@ def load(
 
     y = depth_convert(y, dtype)
     return y, r
+
+
+def info(filepath: str) -> None:
+    raise RuntimeError("No audio I/O backend is available.")
