+++ paddlespeech/audio/utils/__init__.py
@@ -13,11 +13,18 @@
 # limitations under the License.
 from ...cli.utils import DATA_HOME
 from ...cli.utils import MODEL_HOME
+
 from .download import decompress
 from .download import download_and_decompress
 from .download import load_state_dict_from_url
+
 from .error import ParameterError
+
 from .log import Logger
 from .log import logger
+
 from .time import seconds_to_hms
 from .time import Timer
+
+from .numeric import pcm16to32
+from .numeric import depth_convert
