+++ paddlespeech/server/engine/vector/python/vector_engine.py
@@ -17,7 +17,7 @@ from collections import OrderedDict
 import numpy as np
 import paddle
 
-from paddlespeech.audio.backends import load as load_audio
+from paddlespeech.audio import load as load_audio
 from paddlespeech.audio.compliance.librosa import melspectrogram
 from paddlespeech.cli.log import logger
 from paddlespeech.cli.vector.infer import VectorExecutor
